Init: set partitions with MaterializedResult
