+++ daft/runners/partitioning.py
@@ -4,7 +4,7 @@ import sys
 import weakref
 from abc import abstractmethod
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any, Generic, TypeVar
+from typing import TYPE_CHECKING, Any, Generic, Protocol, TypeVar
 from uuid import uuid4
 
 import pyarrow as pa
@@ -92,6 +92,35 @@ class PartitionMetadata(PartialPartitionMetadata):
 PartitionT = TypeVar("PartitionT")
 
 
+class MaterializedResult(Protocol[PartitionT]):
+    """A protocol for accessing the result partition of a PartitionTask.
+
+    Different Runners can fill in their own implementation here.
+    """
+
+    def partition(self) -> PartitionT:
+        """Get the partition of this result."""
+        ...
+
+    def vpartition(self) -> Table:
+        """Get the vPartition of this result."""
+        ...
+
+    def metadata(self) -> PartitionMetadata:
+        """Get the metadata of the partition in this result."""
+        ...
+
+    def cancel(self) -> None:
+        """If possible, cancel execution of this PartitionTask."""
+        ...
+
+    def _noop(self, _: PartitionT) -> None:
+        """Implement this as a no-op.
+        https://peps.python.org/pep-0544/#overriding-inferred-variance-of-protocol-classes
+        """
+        ...
+
+
 class PartitionSet(Generic[PartitionT]):
     def _get_merged_vpartition(self) -> Table:
         raise NotImplementedError()
@@ -126,7 +155,7 @@ class PartitionSet(Generic[PartitionT]):
         raise NotImplementedError()
 
     @abstractmethod
-    def set_partition(self, idx: PartID, part: PartitionT) -> None:
+    def set_partition(self, idx: PartID, part: MaterializedResult[PartitionT]) -> None:
         raise NotImplementedError()
 
     @abstractmethod
